+++ server/routerlicious/packages/services-client/package.json
@@ -56,12 +56,14 @@
     "@types/node": "^10.17.24",
     "axios": "^0.18.0",
     "debug": "^4.1.1",
+    "jwt-decode": "^3.0.0",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
     "@fluidframework/build-common": "^0.19.2",
     "@fluidframework/eslint-config-fluid": "^0.19.1",
     "@types/debug": "^4.1.5",
+    "@types/jwt-decode" : "^2.2.1",
     "@types/mocha": "^5.2.5",
     "@typescript-eslint/eslint-plugin": "~2.17.0",
     "@typescript-eslint/parser": "~2.17.0",
