+++ server/routerlicious/packages/services-client/src/index.ts
@@ -3,6 +3,7 @@
  * Licensed under the MIT License.
  */
 
+export * from "./auth";
 export * from "./gitManager";
 export * from "./historian";
 export * from "./interfaces";
